IMG-198 - updates to modules to account for problems introduced by cluster upgrades
git-svn-id: https://svn.broadinstitute.org/CellProfiler/trunk/CellProfiler@7950 1fc53939-2000-0410-845c-e8453a809027

+++ pyCellProfiler/cellprofiler/cpmath/cpmorphology.py
@@ -791,7 +791,7 @@ def minimum_enclosing_circle(labels, indexes = None,
         #
         min_position = scind.minimum_position(angle_s0vs1,v_labels,
                                               indexes)
-        min_position = fixup_scipy_ndimage_result(min_position)
+        min_position = fixup_scipy_ndimage_result(min_position).astype(int)
         min_position = min_position.flatten()
         #
         # Case 1: minimum angle is obtuse or right. Accept S as the diameter.
